Fix varstore-sb-state exit code
In 'user' mode, varstore-sb-state exit code is the returned value of the
function setup_keys() which returns true on success and false otherwise.
The exit code is then 1 on success and 0 on failure, the opposite of any
command line tool exit code.

This patch fixes the exit code, returning 0 when setup_keys() succeeds
and 1 otherwise.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>