Add support for auth_require_only_pk global var
This variable is used to modify the auth_info struct and set the
'required' field to false for all keys except the PK one. This allows
the setup_keys() function not to fail if only the PK is present and
switch the nvram to user mode.

This will make VMs that are setup with SecureBoot mode on to fail
booting if only the PK is present, which is fine. Otherwise, the nvram
is stucked in setup mode with secure boot disabled, while let you think
the VMs is booted securely.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>