Add get_platform_key() implementation for XAPI backend
This patch implements the backend function get_platform_key(). It
connects to the XAPI and get the key value in the VM platform entry.

The key to get is passed as parameter to get_platform_key() and the
value is returned as an allocated string. The caller is responsible to
free it.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>